Removed agitator and ball spinner
Removed agitator and ball spinner because we took them off the bot. Also
removed the second camera (the one for the rope) because the framerate
was too low to be useful. This is the checkin after the second
competition (Shepherd), so it includes lots of small tweaks to auton
code. Also lowered the speed of the slow-speed mode because it was still
a little too fast.